+++ packages/blade/src/components/Charts/index.ts
@@ -3,3 +3,4 @@ export * from './AreaChart';
 export * from './DonutChart';
 export * from './BarChart';
 export * from './CommonChartComponents';
+export * from './SankeyChart';